# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ashiato API
  version: 0.1.0
  description: Backend-core API for Ashiato.
servers:
  - url: /
...
paths:
...
  /internal/v1/workflow-periods/provision:
    post:
      operationId: provisionWorkflowPeriod
      tags: [internal]
      summary: Provision a workflow period for all organizations
      description: Creates initial meeting (planned) and announcement (draft) entries for the given year/month if they do not exist. Protected by X-Cron-Secret header.
      security:
        - CronSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionWorkflowPeriodRequest"
      responses:
        "200":
          description: Provisioned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionWorkflowPeriodResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BotToken:
      type: apiKey
      in: header
      name: X-Bot-Token
    CronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ProvisionWorkflowPeriodRequest:
      type: object
      required: [year, month]
      properties:
        year:
          type: integer
          minimum: 2000
          maximum: 2100
        month:
          type: integer
          minimum: 1
          maximum: 12
    ProvisionWorkflowPeriodResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...